chore: Cleanup the linking functions
